+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 ### Added
 
+* Add interactive mode for text command
 * Add parameter `history`, `model`, `temp`, `system` for text command
 * Add history with name for text command
+* Add sdk claude for text command
 * Add sdk openai for text command
 * Add localized for `en` and `fr`
 * Read config.ini file